[ refactor ] sessionStorage 관련 코드 custom hook으로 관리
